More test cleanup
